backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI和RAG相关
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# 数据处理
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0

# 向量数据库
chromadb==0.4.18
milvus
neo4j

# 工具库
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 开发工具
black==23.11.0
flake8==6.1.0
